+++ CHANGELOG.md
@@ -226,3 +226,4 @@
 - 2023-12-26: Err lead, tz country, main, find email, +5
 - 2023-12-27: Builder frontier, campaign, geo, chat, +2
 - 2023-12-27: Serde users, main, campaign, qualify, +2
+- 2023-12-28: Wire discover, bettercontact, sender, README, +4